+++ tests/storage/linstor/regular/conftest.py
@@ -1,0 +1,10 @@
+# explicit import for package-scope fixtures
+from tests.storage.linstor.pkgfixtures import (
+    _linstor_config,
+    linstor_redundancy,
+    linstor_sr,
+    lvm_disk_paths,
+    lvm_disks,
+    pool_with_linstor,
+    storage_pool_name,
+)
